アプリ一覧コンポーネントの切り分け
