+++ web/pages/members/kato.tsx
@@ -1,32 +1,15 @@
+import React from 'react';
 import { motion, useMotionValue, useTransform } from 'framer-motion';
 import Social from '../../components/kato/Social';
 import MyImg from "../../components/kato/MyImg";
 import { Transition } from '../../components/kato/transitions';
+import AppsList from '../../components/kato/AppsList';
 
 const KatoPage: React.FC = () => {
   const x = useMotionValue(0);
   const y = useMotionValue(0);
   const rotateX = useTransform(y, [-100, 100], [30, -30]);
   const rotateY = useTransform(x, [-100, 100], [-30, 30]);
-
-  // アプリデータを定義
-  const apps = [
-    {
-      name: "App One",
-      description: "This is a description of App One.",
-      link: "https://example.com/app-one"
-    },
-    {
-      name: "App Two",
-      description: "This is a description of App Two.",
-      link: "https://example.com/app-two"
-    },
-    {
-      name: "App Three",
-      description: "This is a description of App Three.",
-      link: "https://example.com/app-three"
-    }
-  ];
 
   return (
     <motion.section
@@ -98,28 +81,7 @@ const KatoPage: React.FC = () => {
           </div>
         </section>
 
-        <section className="relative z-20 py-12 xl:py-24 text-orange w-full">
-          <div className="container mx-auto backdrop-blur-lg bg-white/20 p-8 rounded-lg shadow-lg">
-            <h2 className="text-4xl font-bold text-center mb-12">My Applications</h2>
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
-              {apps.map((app, index) => (
-                <motion.div
-                  key={index}
-                  className="p-6 bg-blue rounded-lg shadow-lg hover:bg-lightblue transition duration-300"
-                  initial={{ opacity: 0, y: 20 }}
-                  animate={{ opacity: 1, y: 0 }}
-                  transition={{ delay: index * 0.1 }}
-                >
-                  <h3 className="text-2xl font-semibold mb-4">{app.name}</h3>
-                  <p className="mb-4">{app.description}</p>
-                  <a href={app.link} className="text-green-400 hover:underline" target="_blank" rel="noopener noreferrer">
-                    View App
-                  </a>
-                </motion.div>
-              ))}
-            </div>
-          </div>
-        </section>
+        <AppsList />
 
         <video autoPlay loop muted className="absolute top-0 left-0 w-full h-full object-cover">
           <source src="/kato/sky.mp4" type="video/mp4" />
